Separate generator v1 code from v2 code a bit more
As I work through the implementation of v2, I'm finding that the generator code
is specific enough that it's not really helpful to attempt to share the
generator code. I think it will be easier to either copy the v1 code and modify
it, or start from scratch. This will also give me an opportunity to see if I can
improve on it in the process of rewriting it.